02_datalogging/examples/Python/00_Communication/requirements.txt
pyserial
numpy
zmq
